+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This is a small bugfix release.
 ### Changelog
 
 - Upd: Japanese translation
+- Fix: Save playback view #1264
 
 ***