+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This is a small bugfix release.
 - Upd: Japanese translation
 - Fix: Save playback view #1264
 - Fix: JavaScript minification #1264
+- Fix: GPIO set and toggle functions
 
 ***